classifier/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.104.1
filelock==3.13.1
fsspec==2023.10.0
h11==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
Pillow==10.1.0
pydantic==2.5.1
pydantic_core==2.14.3
python-multipart==0.0.6
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
torch==2.1.1
torchvision==0.16.1
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
